Refactor shaping module for better clarity
