+++ test/expected/text-shaping-default.json
@@ -1,5 +1,5 @@
 {
-  "positionedGlyphs": [
+  "placedGlyphs": [
     {
       "codePoint": 97,
       "x": -32.5,
