+++ test/expected/text-shaping-null.json
@@ -1,5 +1,5 @@
 {
-  "positionedGlyphs": [
+  "placedGlyphs": [
     {
       "codePoint": 104,
       "x": -10,
