+++ test/expected/text-shaping-spacing.json
@@ -1,5 +1,5 @@
 {
-  "positionedGlyphs": [
+  "placedGlyphs": [
     {
       "codePoint": 97,
       "x": -38.5,
